pyproject.toml
[tool.poetry]
name = "twitter-media-downloader-webapi"
version = "0.1.0"
description = ""
authors = ["Bakoura Ibrahima <bakouraibrahima@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
python-dotenv = "^1.0.0"
playwright = "^1.40.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
